+++ code/classeine-lib/clsn/ui/HBoxContainer.cpp
@@ -5,15 +5,5 @@ namespace clsn::ui
     HBoxContainer::HBoxContainer()
     : ListContainer("HBoxContainer")
     {
-        initEvents();
-    }
-
-    void HBoxContainer::initEvents()
-    {
-        addActualSizeChangedListener(
-            [this](auto& )
-            {
-                doLayout();
-            });
     }
 }