+++ code/classeine-lib/clsn/ui/VBoxContainer.cpp
@@ -5,15 +5,5 @@ namespace clsn::ui
     VBoxContainer::VBoxContainer()
     : ListContainer("VBoxContainer")
     {
-        initEvents();
-    }
-
-    void VBoxContainer::initEvents()
-    {
-         addActualSizeChangedListener(
-            [this](auto& )
-            {
-                doLayout();
-            });
     }
 }