+++ code/classeine-lib/clsn/ui/XYContainer.cpp
@@ -5,14 +5,5 @@ namespace clsn::ui
     XYContainer::XYContainer()
     : ListContainer("XYContainer")
     {
-        initEvents();
-    }
-
-    void XYContainer::initEvents()
-    {
-        addActualSizeChangedListener([this](auto& )
-        {
-            doLayout();
-        });
     }
 }